move utils function into own folder, update imports
